+++ api/src/main/java/org/grnet/cat/api/endpoints/TemplatesEndpoint.java
@@ -54,11 +54,11 @@ public class TemplatesEndpoint {
 
     @Tag(name = "Template")
     @Operation(
-            summary = "Get assessment template By Actor.",
-            description = "This endpoint retrieves the assessment template corresponding to an actor")
+            summary = "Get assessment template By Type and Actor.",
+            description = "This endpoint retrieves the assessment template corresponding to an actor and a specific type")
     @APIResponse(
             responseCode = "200",
-            description = "Actor's assessment template.",
+            description = "Actor's assessment template by type.",
             content = @Content(schema = @Schema(
                     type = SchemaType.OBJECT,
                     implementation = TemplateDto.class)))
@@ -305,4 +305,58 @@ public class TemplatesEndpoint {
             this.content = content;
         }
     }
+
+    @Tag(name = "Template")
+    @Operation(
+            summary = "Retrieve assessment templates for a specific actor.",
+            description = "This endpoint retrieves the assessment templates for a specific actor." +
+                    "By default, the first page of 10 assessment templates will be returned. You can tune the default values by using the query parameters page and size.")
+    @APIResponse(
+            responseCode = "200",
+            description = "List of assessment templates.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = PageableTemplate.class)))
+    @APIResponse(
+            responseCode = "401",
+            description = "User has not been authenticated.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @APIResponse(
+            responseCode = "403",
+            description = "Not permitted.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @APIResponse(
+            responseCode = "500",
+            description = "Internal Server Error.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @SecurityRequirement(name = "Authentication")
+    @GET
+    @Path("/by-actor/{actor-id}")
+    @Produces(MediaType.APPLICATION_JSON)
+    @Registration
+    public Response getTemplatesByActor(@Parameter(name = "page", in = QUERY,
+            description = "Indicates the page number. Page number must be >= 1.") @DefaultValue("1") @Min(value = 1, message = "Page number must be >= 1.") @QueryParam("page") int page,
+                                       @Parameter(name = "size", in = QUERY,
+                                               description = "The page size.") @DefaultValue("10") @Min(value = 1, message = "Page size must be between 1 and 100.")
+                                       @Max(value = 100, message = "Page size must be between 1 and 100.") @QueryParam("size") int size,
+                                       @Parameter(
+                                               description = "The Actor to retrieve templates.",
+                                               required = true,
+                                               example = "1",
+                                               schema = @Schema(type = SchemaType.NUMBER))
+                                       @PathParam("actor-id") @Valid @NotFoundEntity(repository = ActorRepository.class, message = "There is no Actor with the following id:") Long typeId,
+                                       @Context UriInfo uriInfo) {
+
+        var templates = templateService.getTemplatesByActor(page-1, size, typeId, uriInfo);
+
+        return Response.ok().entity(templates).build();
+    }
+
+
 }